Refactor for poetry and test suites
