+++ quest_test/pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+markers =
+    integration: mark a test as an integration test
+    unit: mark a test as a unit test
