+++ src/quest/__init__.py
@@ -6,8 +6,7 @@ from .wrappers import step, task
 from .external import state, queue, identity_queue, event
 from .historian import Historian
 from .history import History
-from .sql import SQLDatabase, SqlBlobStorage
-from .persistence import LocalFileSystemBlobStorage, PersistentHistory, DynamoDB, DynamoDBBlobStorage, BlobStorage, Blob
+from .persistence import LocalFileSystemBlobStorage, PersistentHistory, BlobStorage, Blob
 from .versioning import version, get_version
 from .manager import WorkflowManager, WorkflowFactory
 from .utils import ainput
@@ -35,20 +34,4 @@ def create_filesystem_manager(
 
     return WorkflowManager(namespace, storage, create_history, factory)
 
-try:
-    from .sql import create_sql_manager
-except ImportError:
-    raise ImportError("quest[sql] is not installed. To use SQL databases, install quest with the sql extra: pip install quest-py[sql]")
 
-def create_dynamodb_manager(
-        namespace: str,
-        factory: WorkflowFactory,
-) -> WorkflowManager:
-    dynamodb = DynamoDB()
-
-    storage = DynamoDBBlobStorage(namespace, dynamodb.get_table())
-
-    def create_history(wid: str) -> History:
-        return PersistentHistory(wid, DynamoDBBlobStorage(wid, dynamodb.get_table()))
-
-    return WorkflowManager(namespace, storage, create_history, factory)
