--- src/quest/storage/__init__.py
@@ -1,3 +1,0 @@
-from .local import LocalFileSystemBlobStorage
-from .blob import BlobStorage, Blob
-
