--- src/quest/storage/local.py
@@ -1,40 +1,0 @@
-import json
-from .blob import BlobStorage, Blob
-from pathlib import Path
-
-class LocalFileSystemBlobStorage(BlobStorage):
-    def __init__(self, root_folder: Path):
-        root_folder.mkdir(parents=True, exist_ok=True)
-        self._root = root_folder
-
-    def _get_file(self, key: str) -> Path:
-        return self._root / (key + '.json')
-
-    def write_blob(self, key: str, blob: Blob):
-        self._get_file(key).write_text(json.dumps(blob))
-
-    def read_blob(self, key: str) -> Blob:
-        return json.loads(self._get_file(key).read_text())
-
-    def has_blob(self, key: str) -> bool:
-        return self._get_file(key).exists()
-
-    def delete_blob(self, key: str):
-        self._get_file(key).unlink()
-
-
-class InMemoryBlobStorage(BlobStorage):
-    def __init__(self):
-        self._data = {}
-
-    def write_blob(self, key: str, blob: Blob):
-        self._data[key] = blob
-
-    def read_blob(self, key: str) -> Blob:
-        return self._data[key]
-
-    def has_blob(self, key: str) -> bool:
-        return key in self._data
-
-    def delete_blob(self, key: str):
-        del self._data[key]
